requirements.txt
aenum==3.1.15
coverage==7.4.3
chromadb==0.4.24
fastapi==0.110.0
langchain==0.1.16
langchain-community==0.0.32
langchain_core==0.1.42
langchain-openai==0.1.3
numpy==1.26.4
openai==1.13.3
pandas==1.5.3
pandasai==2.0.32
pydantic==1.10.13
pytest==8.0.0
python-decouple==3.8
python_dateutil==2.8.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.2
sentence_transformers==2.2.2
torch==2.1.1
wandb==0.16.2
openpyxl==3.1.2